Forms without FORM_TYPE are treated as plain DataForms This commit modifies the DataForm API to only use the FormFieldRegistry when a FORM_TYPE has been set to the DataForm.
If no FORM_TYPE field is set, the form fields are parsed as specified in XEP-0004, and the
field type is determined as follows:
1) the field type set to the field itself
2) the field type set in the <reported> section of result forms
3) the field type default of text-single is used

If a FORM_TYPE field is set, the lookup of form field types looks like this:
1) the field type set to the field itself
2) the field type set in the <reported> section of result forms
3) the FormFieldRegistry is checked for a preregistered type of the field for this FORM_TYPE
4) the field type default of text-single is used

To improve the support for DataForms without a FORM_TYPE, the FilledForm and FillableForm's
can now handle forms without a FORM_TYPE.

In addition a bug was fixed where the FORM_TYPE field was required to be hidden to be
treated as the FORM_TYPE field. This behavior was not correct for submit forms, where
the type of the field may be omitted.