[xdata] Do not lookup fields of forms without FORM_TYPE in registry
The initial implementation was to restrictive, as it would throw
together differents fields, having the same field name but a different
field type. This potentially causes parsing exception, if, e.g. a
boolean value is expected, but a non-boolean string was provided.

We now simply assume that fields are of the default type, if we can
not determine the type. Which is close to the original behavior and
should be, for the most part, correct.